move for loop checks to error checker
